Added Lore of Undeath
I added the lore option to all non-special characters who are wizards. I
didn't add it to special characters, because some or more of them have a
limitation on what lore they use or how they generate/choose spells, or
units like Pink Horrors because they are not wizards instead they are
treated as wizards and I don't know if they are supposed to get access
to this lore. I also didn't add it to Grey Seers because they don't
choose a Lore so I don't know if Lore of Undeath is supposed to come as
a third lore from which they can generate spells or is it supposed to
replace Ruin or Plague or is it either all spells from Undeath or
normally or something else.